sql: refactoring to use the new package
